working light and deep sleep, tested on ESP
