+++ dependencies.lock
@@ -2,7 +2,7 @@ dependencies:
   idf:
     source:
       type: idf
-    version: 5.5.1
+    version: 5.4.0
 manifest_hash: d93cf42f7d1a782374d661974f5fc42f717972fb336e163601f276be74b7b9d5
 target: esp32
 version: 2.0.0
